Move schemeshard index-related code to index/
