+++ ydb/core/tx/schemeshard/index/build_index.cpp
@@ -1,7 +1,7 @@
-#include "schemeshard_build_index.h"
+#include <ydb/core/tx/schemeshard/index/build_index.h>
 
-#include "schemeshard_impl.h"
-#include "schemeshard_index_build_info.h"
+#include <ydb/core/tx/schemeshard/schemeshard_impl.h>
+#include <ydb/core/tx/schemeshard/index/index_build_info.h>
 #include <ydb/core/protos/flat_scheme_op.pb.h>
 
 namespace NKikimr {
