+++ ydb/core/tx/schemeshard/index/build_index_helpers.h
