+++ ydb/core/tx/schemeshard/index/build_index_tx_base.h
@@ -1,7 +1,7 @@
 #pragma once
 
-#include "schemeshard_impl.h"
-#include "schemeshard_index_build_info.h"
+#include <ydb/core/tx/schemeshard/schemeshard_impl.h>
+#include <ydb/core/tx/schemeshard/index/index_build_info.h>
 
 #include <ydb/public/api/protos/ydb_issue_message.pb.h>
 #include <ydb/public/api/protos/ydb_status_codes.pb.h>
