+++ ydb/core/tx/schemeshard/index/operation_create_indexed_table.cpp
@@ -1,7 +1,7 @@
-#include "schemeshard__op_traits.h"
-#include "schemeshard__operation_common.h"
-#include "schemeshard__operation_part.h"
-#include "schemeshard_index_utils.h"
+#include <ydb/core/tx/schemeshard/schemeshard__op_traits.h>
+#include <ydb/core/tx/schemeshard/schemeshard__operation_common.h>
+#include <ydb/core/tx/schemeshard/schemeshard__operation_part.h>
+#include <ydb/core/tx/schemeshard/index/index_utils.h>
 
 #include <ydb/core/protos/flat_scheme_op.pb.h>
 #include <ydb/core/protos/flat_tx_scheme.pb.h>
