+++ ydb/core/tx/schemeshard/index/operation_drop_indexed_table.cpp
@@ -1,7 +1,7 @@
-#include "schemeshard__operation_common.h"
-#include "schemeshard__operation_part.h"
-#include "schemeshard_impl.h"
-#include "schemeshard_path_element.h"
+#include <ydb/core/tx/schemeshard/schemeshard__operation_common.h>
+#include <ydb/core/tx/schemeshard/schemeshard__operation_part.h>
+#include <ydb/core/tx/schemeshard/schemeshard_impl.h>
+#include <ydb/core/tx/schemeshard/schemeshard_path_element.h>
 
 #include <ydb/core/base/path.h>
 #include <ydb/core/base/table_index.h>
