+++ ydb/core/tx/schemeshard/index/operation_move_index.cpp
@@ -1,8 +1,8 @@
-#include "schemeshard__operation_common.h"
-#include "schemeshard__operation_part.h"
-#include "schemeshard_cdc_stream_common.h"
-#include "schemeshard_impl.h"
-#include "schemeshard_path_element.h"
+#include <ydb/core/tx/schemeshard/schemeshard__operation_common.h>
+#include <ydb/core/tx/schemeshard/schemeshard__operation_part.h>
+#include <ydb/core/tx/schemeshard/schemeshard_cdc_stream_common.h>
+#include <ydb/core/tx/schemeshard/schemeshard_impl.h>
+#include <ydb/core/tx/schemeshard/schemeshard_path_element.h>
 
 #include <ydb/core/base/path.h>
 #include <ydb/core/mind/hive/hive.h>
