+++ ydb/core/tx/schemeshard/schemeshard__notify.cpp
@@ -1,7 +1,7 @@
 #include "schemeshard_impl.h"
-#include "schemeshard_index_build_info.h"
 
 #include <ydb/core/base/appdata.h>
+#include <ydb/core/tx/schemeshard/index/index_build_info.h>
 
 namespace NKikimr {
 namespace NSchemeShard {
