+++ ydb/core/tx/schemeshard/schemeshard_import.cpp
@@ -1,9 +1,10 @@
 #include "schemeshard_import.h"
 
 #include "schemeshard_impl.h"
-#include "schemeshard_index_build_info.h"
 #include "schemeshard_import_getters.h"
 #include "schemeshard_import_helpers.h"
+
+#include <ydb/core/tx/schemeshard/index/index_build_info.h>
 
 #include <util/generic/xrange.h>
